+++ flumine/execution/betfairexecution.py
@@ -185,7 +185,7 @@ class BetfairExecution(BaseExecution):
             session=session,
         )
 
-    def _execution_helper(  # todo retry!
+    def _execution_helper(
         self,
         trading_function: Callable,
         order_package: BaseOrderPackage,
@@ -202,8 +202,13 @@ class BetfairExecution(BaseExecution):
                         "response": e,
                         "order_package": order_package.info,
                     },
+                    exc_info=True,
                 )
-                return  # session not returned on purpose
+                if order_package.retry():
+                    self.handler_queue.put(order_package)
+
+                self._return_http_session(http_session, err=True)
+                return
             logger.info(
                 "execute_%s" % trading_function.__name__,
                 extra={