+++ tests/test_execution.py
@@ -104,13 +104,26 @@ class BaseExecutionTest(unittest.TestCase):
         self.assertEqual(self.execution._get_http_session(), 2)
         self.assertEqual(self.execution._get_http_session(), 1)
         self.assertEqual(self.execution._get_http_session(), mock_requests.Session())
+        self.assertEqual(self.execution._sessions_created, 1)
 
     def test__return_http_session(self):
-        self.execution._return_http_session(1)
-        self.assertEqual(self.execution._sessions, [1])
-        self.execution._return_http_session(2)
-        self.execution._return_http_session(3)
-        self.assertEqual(self.execution._sessions, [1, 2])
+        mock_session = mock.Mock()
+        self.execution._return_http_session(mock_session)
+        self.assertEqual(self.execution._sessions, [mock_session])
+        self.execution._return_http_session(mock_session)
+        self.execution._return_http_session(mock_session)
+        self.assertEqual(self.execution._sessions, [mock_session, mock_session])
+
+    def test__return_http_session_close(self):
+        self.execution._sessions = [1, 2]
+        mock_session = mock.Mock()
+        self.execution._return_http_session(mock_session)
+        mock_session.close.assert_called_with()
+
+    def test__return_http_session_err_close(self):
+        mock_session = mock.Mock()
+        self.execution._return_http_session(mock_session, err=True)
+        mock_session.close.assert_called_with()
 
     @mock.patch("flumine.execution.baseexecution.OrderEvent")
     def test__order_logger_place(self, mock_order_event):
@@ -658,23 +671,51 @@ class BetfairExecutionTest(unittest.TestCase):
         mock_trading_function.assert_not_called()
         mock__return_http_session.assert_called_with(mock_session)
 
+    @mock.patch("flumine.execution.betfairexecution.BetfairExecution.handler_queue")
     @mock.patch(
         "flumine.execution.betfairexecution.BetfairExecution._return_http_session"
     )
-    def test__execution_helper_error(self, mock__return_http_session):
+    def test__execution_helper_error(
+        self, mock__return_http_session, mock_handler_queue
+    ):
         mock_trading_function = mock.Mock()
         mock_trading_function.__name__ = "test"
         mock_trading_function.side_effect = BetfairError()
         mock_session = mock.Mock()
         mock_order_package = mock.Mock()
         mock_order_package.info = {}
+        mock_order_package.retry.return_value = True
         self.assertIsNone(
             self.execution._execution_helper(
                 mock_trading_function, mock_order_package, mock_session
             )
         )
         mock_trading_function.assert_called_with(mock_order_package, mock_session)
-        mock__return_http_session.assert_not_called()
+        mock__return_http_session.assert_called_with(mock_session, err=True)
+        mock_handler_queue.put.assert_called_with(mock_order_package)
+
+    @mock.patch("flumine.execution.betfairexecution.BetfairExecution.handler_queue")
+    @mock.patch(
+        "flumine.execution.betfairexecution.BetfairExecution._return_http_session"
+    )
+    def test__execution_helper_error_no_retry(
+        self, mock__return_http_session, mock_handler_queue
+    ):
+        mock_trading_function = mock.Mock()
+        mock_trading_function.__name__ = "test"
+        mock_trading_function.side_effect = BetfairError()
+        mock_session = mock.Mock()
+        mock_order_package = mock.Mock()
+        mock_order_package.info = {}
+        mock_order_package.retry.return_value = False
+        self.assertIsNone(
+            self.execution._execution_helper(
+                mock_trading_function, mock_order_package, mock_session
+            )
+        )
+        mock_trading_function.assert_called_with(mock_order_package, mock_session)
+        mock__return_http_session.assert_called_with(mock_session, err=True)
+        mock_handler_queue.put.assert_not_called()
 
 
 class SimulatedExecutionTest(unittest.TestCase):